pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "baseball-manager"
version = "0.1.0"
description = "Yahoo Fantasy Baseball AI Manager"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "requests-oauthlib>=1.3.1",
    "xmltodict>=0.13.0",
    "pandas>=2.0.0",
    "tabulate>=0.9.0",
    "schedule>=1.2.0",
    "click>=8.1.0",
]

[project.scripts]
bbm = "baseball_manager.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
